move poisson_xover into its own header
