+++ fwdpp/Makefile.in
@@ -298,6 +298,7 @@ pdfdir = @pdfdir@
 prefix = @prefix@
 program_transform_name = @program_transform_name@
 psdir = @psdir@
+runstatedir = @runstatedir@
 sbindir = @sbindir@
 sharedstatedir = @sharedstatedir@
 srcdir = @srcdir@
@@ -317,6 +318,7 @@ pkginclude_HEADERS = debug.hpp \
 	IO.tcc \
 	mutation.hpp \
 	mutation.tcc \
+	poisson_xover.hpp \
 	recombination.hpp \
 	recombination.tcc \
 	sample_diploid.hpp \
