+++ testsuite/fixtures/sugar_fixtures.hpp
@@ -7,6 +7,7 @@
 #include <fwdpp/sugar/multiloc.hpp>
 #include <fwdpp/sugar/infsites.hpp>
 #include <fwdpp/sugar/GSLrng_t.hpp>
+#include <fwdpp/poisson_xover.hpp>
 #include <fwdpp/recombination.hpp>
 #include <fwdpp/fitness_models.hpp>
 #include <fwdpp/extensions/regions.hpp>
